+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 'View Logs' tree view now accessible for CacheProgress objects
 - Added query/result tabs to CLI GUI Sql editor
 - Console GUI now shows important information (e.g. 'Disabled') in brackets next to items where state is highly important
+- Added new command RunSupportingSql
 - Console GUI root nodes now offer sensible commands (e.g. create new Catalogue)
 - Added Value column to tree views (allows user to quickly see current arguments' values)
 - Added 'other' checkbox to 'Create Catalogue by importing a file' (for selecting custom piplelines)